+++ include/ecgen/combin_old.hpp
@@ -7,23 +7,32 @@
 namespace ecgen {
 
 /**
- * @brief emk (GEN)
+ * Generates all k-combinations from a set of n elements as pairs of indices.
  *
- * @param[in] n
- * @param[in] k
- * @return cppcoro::recursive_generator<std::pair<size_t, size_t>>
+ * This is a recursive generator function that yields all possible
+ * k-combinations of integer indices ranging from 0 to n-1.
+ *
+ * @param[in] n - The size of the set to generate combinations from.
+ * @param[in] k - The size of each generated combination.
+ * @returns A recursive generator yielding all k-combinations as pairs of
+ * indices.
  */
-extern auto emk_gen(size_t n, size_t k)
-    -> cppcoro::recursive_generator<std::pair<size_t, size_t>>;
+extern auto emk_gen(int n, int k)
+    -> cppcoro::recursive_generator<std::pair<int, int>>;
 
 /**
- * @brief emk (NEG)
+ * Generates all k-combinations from a set of n elements, without repetition.
  *
- * @param[in] n
- * @param[in] k
- * @return cppcoro::recursive_generator<std::pair<size_t, size_t>>
+ * This is a recursive generator function that yields all possible
+ * k-combinations of integer indices ranging from 0 to n-1, without repeating
+ * indices.
+ *
+ * @param[in] n - The size of the set to generate combinations from.
+ * @param[in] k - The size of each generated combination.
+ * @returns A recursive generator yielding all k-combinations as pairs of
+ * indices.
  */
-extern auto emk_neg(size_t n, size_t k)
-    -> cppcoro::recursive_generator<std::pair<size_t, size_t>>;
+extern auto emk_neg(int n, int k)
+    -> cppcoro::recursive_generator<std::pair<int, int>>;
 
 } // namespace ecgen
